remove top-level gruntFinishTask function
go back to threading the callback through the other async functions. this revealed some bugs and opportunities for improvement
saveMedia doesn't have to save the json file. let the calling function do that, and let saveMedia do one thing
the batch unzipper spawns a bunch of async operations, which need a counter, although what it really needs is to be broken way down