clarify working of getMediaAssets function
this function previously returned one result with side-effects on one of its arguments. now, both effects of the function are returned in an object with two properties. the obj argument is no longer rewritten, instead a modified deep copy is made. this is expressed in terms of a mapping function which is passed to a generic 'transformJson' function, which acts like an Array.map that can recurse into nested Object/Array structures.