clarify behavior of mediaOut and mediaCwd
previously mediaOut and mediaCwd defaulted to null. if mediaOut was null or undefined, this would prevent it from downloading media.
turns out that mediaCwd (at least) cannot be null, as its value needs to be known when getMediaAssets rewrites URLs in the JSON.
now both mediaOut and mediaCwd default to the current directory (not sure if this is a good idea but it's made explicit at least). Media download is also specified explicitly now, with a 'downloadMedia' option

downloadMedia is introduced in a backwards-compatible way. if mediaOut is unset, it defaults to false. otherwise it defaults to true.