fix downloaded media truncation
Event listeners were wrong. Instead of listening for the 'end' event on
the input stream, should have been listening for 'finish' on
the output stream. This is dispatched after the output stream is flushed
and closed. It's only been working thus far through luck - but a
computer under heavy utilization revealed that the file was being moved
out of the temp folder too soon, if it was listening for the 'end'
event.